requirements.txt
pyTelegramBotAPI==4.26.0
schedule==1.2.2
SQLAlchemy==2.0.38
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.6
